package/dbusbroker: new package
dbus-broker is an alternate implementation of a dbus dameon. It can be
used as a drop-in replacement for the system bus daemon, as well as the
session bus daemon.

dbus-broker is (basically, and as far as we're concerned in Buildroot)
split in two components:

  - the actual message bus daemon, that relays messages across clients

  - a launcher, which is responsible for setting various aspects of the
    bus, like setting the policy et al. and opening the socket(s) the
    message bus daemon will have to listen on...

The launcher can only be used in a systemd setup (it makes heavy use of
systemd facilities), while the message bus is generic. However, the
message bus daemon is useless without a launcher. There does not exist a
non-systemd launcher, which makes dbus-broker actually a systemd-only
package; this can be revisited when/if a non-systemd launcher appears.

There are two cases:

 1. original dbus disabled

    Here, we install the config files and systemd socket activation
    units; dbus-broker provides the system and sessions bus daemons.

 2. original dbus enabled

    In this case, we do not install the config files and systemd socket
    activation units: they all are provided by the original dbus.

    In this situation, the default system and sessions message bus are
    the original dbus; dbus-broker is not enabled.

    However, users may opt-in to use dbus-broker in a few ways:
      - at build-time: provide drop-in units in an overlay;
      - at build-time: call systemctl enable/disable from a post-build
        script;
      - at runtime (on a RW filesystem): by calling systemctl
        enable/disable

Note that, contrary to the original dbus, we do not define a user to run
the system bus as non-root; this is handled by systemd.

As for the licensing terms: they are pretty trivial for dbus-broker
itself, but it makes use of third-party code that it inherits as git
submodules (that are bundeld in the release archive). Thus the licensing
is a bit convoluted... The third-party codes claim to be licensed as
"Apache-2.0 and LGP-2.1+" in their AUTHORS files, but at the same time
claim "**Apache-2.0** OR **LGPL-2.1-or-later**" in their README files.
The individual source files (that are used) do not seem to have any
licensing header to clarify the situation.

Signed-off-by: Norbert Lange <nolange79@gmail.com>
[yann.morin.1998@free.fr:
  - don't select systemd; depend on it instead
  - only install config files and systemd units wihtout original dbus
  - fix licensing info
  - entirely reword and extend the commit log
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>

---
Changes v3 -> v4  (Yann, respiining after review by Norbert):
  - drop the non-systemd case
  - drop the launcher option
  - reinstate BR2_COREUTILS_HOST_DEPENDENCY and ln --relative

Changes v2 -> v3  (Norbert, respinning after Yann):
  - add an own config entry for dbus-broker-launch
    enabled by default if systemd init is used
  - undo BR2_COREUTILS_HOST_DEPENDENCY
  - undo adding dbus user - never used by this package
  - add condtional audit dependency
  - cleanup conditional logic a bit

Changes v1 -> v2 (Yann):
  - make launcher conditional
  - don't select systemd; don't depend on it either
  - don't install systemd units without systemd
  - only install config files and systemd units wihtout original dbus
  - rename hooks with meaningful names
  - fix licensing info
  - entirely reword and extend the commit log