support/run-test: add test for systemd using dbus-broker
Add four new tests for systemd (rw and ro in each case):
  - use dbus-broker instead of the original dbus
  - use the original dbus, with dbus-broker installed

The first two extend the existing TestInitSystemSystemdRoIfupdown test,
by just enabling dbus-broker; the second ones extend this further, by
explicitly enabling the original dbus.

Presence of the original dbus and dbus-broker on the same system is
valid: the original dbus is used as the system bus daemon, while
dbus-broker is used for (some of) the session bus(es). We do not test
the session bus for now, but the new test ensures the system bus daemon
is not impacted by the presence of dbus-broker.

Note: the 'full' test-case enables all systemd options, and some of them
do pull the original dbus package, so we can't use that to test the
integration of dbus-broker; instead, we extend the ifupdown case, which
does not enable the original dbus.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Norbert Lange <nolange79@gmail.com>